update counts after QT scoring
